Add shim xdg-open implementation
As many 3rd party applications rely on xdg-utils for interacting
with the desktop environment, it makes sense for runtimes to provide
the expected tools. However there's no good reason to ship the usual
shell scripts, as the desktop abstraction is already provided by the
portal layer when running in a sandbox. Just add a small command line
tool that directly calls out to the portal.