Add Travis-CI integration
This is similar to what's in xdg-dbus-proxy, which in turn is based on
dbus. The scripts in tests/ are reasonably CI-system-agnostic and should
be feasible to hook up to Gitlab-CI or similar if desired (as has been
done in dbus). It should also be reasonably straightforward to build
in a non-Debian-derived container by adding an alternative code path
alongside the various apt-get calls.

If the baseline distribution is meant to be older or newer than Ubuntu
14.04 LTS, we can adjust the suites used: Debian releases and Ubuntu
LTS releases are about a year apart, so we can easily hit about 50%
of possible GLib branches.

Signed-off-by: Simon McVittie <smcv@collabora.com>