xdg-email, xdg-open: Terminate GOptionEntry array correctly
g_option_context_add_main_entries() is documented to take an array of
GOptionEntry, terminated by a NULL-filled entry; however, these tools
did not terminate the array.

In practice it appears that current toolchains place enough zero bytes
after the array that this worked as intended on most architectures, but
this is not the case on the Debian mips64el and riscv64 ports.

Signed-off-by: Simon McVittie <smcv@collabora.com>