xdg-open: Avoid the use of O_PATH fds
As of xdg-desktop-portal 1.8.0, the writable checks become far more
strict, meaning that passing an O_PATH fd that the current application
has write access to without declaring that it should be writable will
result in the open failing. This drops O_PATH in favor of the new
standard fd support, and opens the descriptor as read-only to match
the GLib behavior.

Fixes flatpak/flatpak#3879.