flatpak-spawn: Add --share-pids option
This option is similar to --expose-pids, except that instead of making
the subsandbox use a nested pid namespace inside the parent's, it makes
the subsandbox share the parent's pid namespace as-is, so that process
IDs in the parent and the subsandbox are interchangeable. This will
be useful if the parent and the subsandbox communicate via protocols
that assume a global view of the process ID namespace, for example
passing process IDs across an AF_UNIX socket or in shared memory.

See https://github.com/flatpak/flatpak/pull/4060 for the server side of
this feature.

Signed-off-by: Simon McVittie <smcv@collabora.com>