flatpak-spawn: Validate --forward-fd arguments early
This avoids accidentally forwarding a fd that flatpak-spawn opened for
its own purposes, like the signalfd for forward_signals(), which would
result in confusing error messages when flatpak-spawn tries to work
with a fd that has already been closed.

A frequent symptom of this is that apps that use a sandbox-escape via
`flatpak-spawn --host` to mount FUSE filesystems hard-code that the
`$_FUSE_COMMFD` is always going to be 3, which is not actually true
(flatpak/flatpak-xdg-utils#46).

Signed-off-by: Simon McVittie <smcv@collabora.com>